# openapi.yaml
openapi: 3.0.3
info:
  title: University Schedule Management API
  version: '1.3'
  description: >
    API for managing university schedules, classes, user registration, and schedule generation on a weekly basis per semester.
servers:
  - url: https://api.example.com
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Classroom:
      type: object
      required:
        - id
        - name
        - capacity
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the classroom
        name:
          type: string
          description: Name of the classroom
        capacity:
          type: integer
          description: Maximum number of students the classroom can accommodate
        created_at:
          type: string
          format: date-time
          description: Timestamp when the classroom was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the classroom was last updated
      example:
        id: 1
        name: "Physics 101"
        capacity: 30
        created_at: "2024-10-26T12:34:56.789Z"
        updated_at: "2024-10-26T12:34:56.789Z"
    Semester:
      type: object
      required:
        - id
        - title
        - start_date
        - end_date
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the semester
        title:
          type: string
          description: Title of the semester
        start_date:
          type: string
          format: date
          description: Start date of the semester
        end_date:
          type: string
          format: date
          description: End date of the semester
        created_at:
          type: string
          format: date-time
          description: Timestamp when the semester was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the semester was last updated
      example:
        id: "1"
        title: "Fall 2024"
        start_date: "2024-09-01"
        end_date: "2024-12-31"
        created_at: "2024-10-26T12:34:56.789Z"
        updated_at: "2024-10-26T12:34:56.789Z"
    CreateSemesterDto:
      type: object
      required:
        - title
        - start_date
        - end_date
      properties:
        title:
          type: string
          description: Title of the semester
        start_date:
          type: string
          format: date
          description: Start date of the semester in ISO format
        end_date:
          type: string
          format: date
          description: End date of the semester in ISO format
      example:
        title: "Spring 2025"
        start_date: "2025-01-15"
        end_date: "2025-05-15"
    UpdateSemesterDto:
      type: object
      properties:
        title:
          type: string
          description: Title of the semester
        start_date:
          type: string
          format: date
          description: Start date of the semester in ISO format
        end_date:
          type: string
          format: date
          description: End date of the semester in ISO format
      example:
        title: "Advanced Spring 2025"
        start_date: "2025-01-20"
        end_date: "2025-05-20"
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Unauthorized access"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Resource not found"
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: "Invalid input data"
security:
  - BearerAuth: []
paths:
...
  /v1/semesters:
    post:
      tags:
        - Semesters
      summary: Create a new semester
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSemesterDto'
            example:
              title: "Spring 2025"
              start_date: "2025-01-15"
              end_date: "2025-05-15"
      responses:
        '201':
          description: Semester created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Semester'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      tags:
        - Semesters
      summary: Retrieve all semesters
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of semesters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Semester'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/semesters/{semesterId}:
    get:
      tags:
        - Semesters
      summary: Retrieve a semester by ID
      security:
        - BearerAuth: []
      parameters:
        - name: semesterId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the semester to retrieve
      responses:
        '200':
          description: Semester details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Semester'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Semesters
      summary: Update a semester by ID
      security:
        - BearerAuth: []
      parameters:
        - name: semesterId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the semester to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSemesterDto'
            example:
              title: "Advanced Spring 2025"
              start_date: "2025-01-20"
              end_date: "2025-05-20"
      responses:
        '200':
          description: Semester updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Semester'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Semesters
      summary: Delete a semester by ID
      security:
        - BearerAuth: []
      parameters:
        - name: semesterId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the semester to delete
      responses:
        '204':
          description: Semester deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Existing Endpoints for Classrooms
  /v1/classrooms:
    get:
      tags:
        - Classrooms
      summary: Get all classrooms
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of classrooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Classroom'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Classrooms
      summary: Create a new classroom
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - capacity
              properties:
                name:
                  type: string
                  description: Name of the classroom
                capacity:
                  type: integer
                  description: Maximum number of students the classroom can accommodate
              example:
                name: "Chemistry 201"
                capacity: 25
      responses:
        '201':
          description: Classroom created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classroom'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/classrooms/{classroomId}:
    get:
      tags:
        - Classrooms
      summary: Get a classroom by ID
      security:
        - BearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the classroom to retrieve
      responses:
        '200':
          description: Classroom details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classroom'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Classrooms
      summary: Update a classroom
      security:
        - BearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the classroom to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the classroom
                capacity:
                  type: integer
                  description: Maximum number of students the classroom can accommodate
              example:
                name: "Advanced Physics 301"
                capacity: 35
      responses:
        '200':
          description: Classroom updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classroom'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Classrooms
      summary: Delete a classroom
      security:
        - BearerAuth: []
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the classroom to delete
      responses:
        '204':
          description: Classroom deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'